refactor: rename layout.rs to guide.rs which provides a guide component, not actually a layout
