+++ src/web/components/guides/mod.rs
@@ -1,6 +1,6 @@
 pub mod author;
 pub mod autumn;
 pub mod category;
-pub mod layout;
+pub mod guide;
 pub mod outline;
 pub mod sidebar;
